FX-3130 Repositioned fair organizer name and period
